requirements.txt
async-timeout==4.0.2
blinker==1.6.2
boto3==1.26.148
botocore==1.29.148
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.7
Flask==2.3.2
Flask-Cors==3.0.10
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
Pillow==9.5.0
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==4.5.5
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.3.4
WTForms==3.0.1
